+++ inc/le_Counter.hpp
@@ -27,14 +27,14 @@ public:
      * @param e The element to connect from.
      * @param outputSlot The output slot of the element to connect from.
      */
-    void SetInput_CountUp(le_Base<bool>* e, uint16_t outputSlot);
+    void SetInput_CountUp(le_Base<bool>* e, uint8_t outputSlot);
 
     /**
      * @brief Sets the input for resetting the counter.
      * @param e The element to connect from.
      * @param outputSlot The output slot of the element to connect from.
      */
-    void SetInput_Reset(le_Base<bool>* e, uint16_t outputSlot);
+    void SetInput_Reset(le_Base<bool>* e, uint8_t outputSlot);
 
 private:
     uint16_t uCountFinal;  ///< The final count value.
