+++ src/le_AND.cpp
@@ -4,7 +4,7 @@
  * @brief Constructor implementation that initializes the AND element with a specified number of inputs.
  * @param nInputs Number of inputs for the AND element.
  */
-le_AND::le_AND(uint16_t nInputs) : le_Base<bool>(nInputs, 1)
+le_AND::le_AND(uint8_t nInputs) : le_Base<bool>(nInputs, 1)
 {
     // Do nothing...
 }
@@ -19,7 +19,7 @@ void le_AND::Update(float timeStep)
     bool nextValue = true;
 
     // Iterate through all input values and apply logical AND
-    for (uint16_t i = 0; i < this->nInputs; i++)
+    for (uint8_t i = 0; i < this->nInputs; i++)
     {
         le_Base<bool>* e = (le_Base<bool>*)(this->_inputs[i]);
         if (e != nullptr)
@@ -39,7 +39,7 @@ void le_AND::Update(float timeStep)
  * @param outputSlot The output slot of the element to connect from.
  * @param inputSlot The input slot of this AND element to connect to.
  */
-void le_AND::SetInput(le_Base<bool>* e, uint16_t outputSlot, uint16_t inputSlot)
+void le_AND::SetInput(le_Base<bool>* e, uint8_t outputSlot, uint8_t inputSlot)
 {
     // Use the base class connection function
     le_Base::Connect(e, outputSlot, this, inputSlot);
