+++ src/le_Analog1PWinding.cpp
@@ -62,7 +62,7 @@ void le_Analog1PWinding::Update(float timeStep)
  * @param e The element providing the input value for Winding.
  * @param outputSlot The output slot of the element providing the input.
  */
-void le_Analog1PWinding::SetInput_Winding(le_Base<float>* e, uint16_t outputSlot)
+void le_Analog1PWinding::SetInput_Winding(le_Base<float>* e, uint8_t outputSlot)
 {
     le_Element::Connect(e, outputSlot, this, 0);
 }
@@ -72,7 +72,7 @@ void le_Analog1PWinding::SetInput_Winding(le_Base<float>* e, uint16_t outputSlot
  * @param e The element providing the reference real input value.
  * @param outputSlot The output slot of the element providing the reference real input.
  */
-void le_Analog1PWinding::SetInput_RefReal(le_Base<float>* e, uint16_t outputSlot)
+void le_Analog1PWinding::SetInput_RefReal(le_Base<float>* e, uint8_t outputSlot)
 {
     le_Element::Connect(e, outputSlot, this, 1);
 }
@@ -82,7 +82,7 @@ void le_Analog1PWinding::SetInput_RefReal(le_Base<float>* e, uint16_t outputSlot
  * @param e The element providing the reference imaginary input value.
  * @param outputSlot The output slot of the element providing the reference imaginary input.
  */
-void le_Analog1PWinding::SetInput_RefImag(le_Base<float>* e, uint16_t outputSlot)
+void le_Analog1PWinding::SetInput_RefImag(le_Base<float>* e, uint8_t outputSlot)
 {
     le_Element::Connect(e, outputSlot, this, 2);
 }
