+++ src/le_Element.cpp
@@ -4,17 +4,17 @@
  * @brief Constructor that initializes the element with a specified number of inputs.
  * @param nInputs Number of inputs for the element.
  */
-le_Element::le_Element(uint16_t nInputs)
+le_Element::le_Element(uint8_t nInputs)
 {
     // Set the number of inputs
     this->nInputs = nInputs;
 
     // Allocate memory for inputs and output slots
     this->_inputs = new le_Element * [nInputs];
-    this->_outputSlots = new uint16_t[nInputs];
+    this->_outputSlots = new uint8_t[nInputs];
 
     // Initialize input references and output slots
-    for (uint16_t i = 0; i < nInputs; i++)
+    for (uint8_t i = 0; i < nInputs; i++)
     {
         this->_outputSlots[i] = 0;
         this->_inputs[i] = nullptr;
@@ -74,7 +74,7 @@ void le_Element::Update(float timeStamp)
  * @param input The input element.
  * @param inputSlot The slot of the input element.
  */
-void le_Element::Connect(le_Element* output, uint16_t outputSlot, le_Element* input, uint16_t inputSlot)
+void le_Element::Connect(le_Element* output, uint8_t outputSlot, le_Element* input, uint8_t inputSlot)
 {
     // Connect the output element to the input element
     input->_inputs[inputSlot] = output;
